refactor: move device status into wire identity module
It isn't used elsewhere, so we may as well just define it inline
